+++ lib/models/money_objects/securities/securities.dart
@@ -1,6 +1,8 @@
 import 'package:money/helpers/json_helper.dart';
+import 'package:money/models/money_objects/investments/investments.dart';
 import 'package:money/models/money_objects/money_objects.dart';
 import 'package:money/models/money_objects/securities/security.dart';
+import 'package:money/storage/data/data.dart';
 
 // Exports
 export 'package:money/models/money_objects/securities/security.dart';
@@ -23,6 +25,15 @@ class Securities extends MoneyObjects<Security> {
     clear();
     for (final MyJson row in rows) {
       appendMoneyObject(Security.fromJson(row));
+    }
+  }
+
+  @override
+  void onAllDataLoaded() {
+    for (final Security security in iterableList()) {
+      final List<Investment> list = Investments.getInvestmentsFromSecurity(security.uniqueId);
+      security.counts.value = list.length;
+      security.balance.value.amount = Investments.getProfit(list);
     }
   }